% Generated by roxygen2 (4.0.1): do not edit by hand
\name{call_frog}
\alias{call_frog}
\title{Call a frog (Dutch lemmatizer and dependency parser) instance running in daemon mode}
\usage{
call_frog(text, host = "localhost", port = 9772)
}
\arguments{
\item{text:}{The text to parse, may contain}

\item{terms:}{a vector of terms of length equal to docids}

\item{freqs:}{an optional vector giving the frequency of terms}
}
\value{
an object of type DocumentTermMatrix (from the tm package)
}
\description{
See http://ilk.uvt.nl/frog/
To install frog and run as daemon (assuming debian/ubuntu), run:
$ sudo apt-get install frog frogdata ucto
$ frog -S 9772
}
\details{
Note that if something is wrong, it is quite possible that this call will totall hang
your R session as it is waiting for output on the socket, so use with caution!
}

