% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_dtm}
\alias{create_dtm}
\title{Create a document term matrix from a token list}
\usage{
create_dtm(docs, terms, freqs = rep(1, length(terms)), weighting = weightTf)
}
\arguments{
\item{docs:}{a vector that identifies to which document a token belongs}

\item{terms:}{a vector of terms of length equal to docs}

\item{freqs:}{an optional vector giving the frequency of each term}

\item{weighting:}{the optional weighting for tm (default: term frequency)}
}
\value{
an object of type DocumentTermMatrix (from the tm package)
}
\description{
Create a document term matrix from a token list
}

