% Generated by roxygen2 (4.0.1): do not edit by hand
\name{call_frog}
\alias{call_frog}
\title{Call a frog (Dutch lemmatizer and dependency parser) instance running in daemon mode}
\usage{
call_frog(text, host = "localhost", port = 9772, verbose = T)
}
\arguments{
\item{text:}{The text(s) to parse}

\item{host:}{The hostname for the frog server}

\item{port:}{The port the frog server is listening on}

\item{verbose:}{If true, output a message for each document}
}
\value{
a data frame of tokens with columns for lemma, pos, etc
}
\description{
See http://ilk.uvt.nl/frog/
To install frog and run as daemon (assuming debian/ubuntu), run:
$ sudo apt-get install frog frogdata ucto
$ frog -S 9772
}
\details{
A separate call to frog is made for each text in the text input vector.

Note that if something is wrong, it is quite possible that this function will totall hang
your R session as it is waiting for output on the socket, so use with caution!
}

